requirements.txt
aiogram
asyncio
dataclasses
logging
python-dotenv
psycopg2